+++ hew-codegen/src/mlir/MLIRGen.cpp
@@ -5022,724 +5022,6 @@ void MLIRGen::generateTraitDefaultMethod(const ast::TraitMethod &method,
 }
 
 // ============================================================================
-// Function generation
-// ============================================================================
-
-mlir::func::FuncOp MLIRGen::generateFunction(const ast::FnDecl &fn, const std::string &nameOverride,
-                                             std::optional<mlir::Location> fallbackLoc) {
-  // Create symbol table scopes for this function
-  SymbolTableScopeT varScope(symbolTable);
-  MutableTableScopeT mutScope(mutableVars);
-
-  // Determine parameter types
-  auto typeLoc = [&](const ast::Spanned<ast::TypeExpr> &type) {
-    return type.span.end > type.span.start ? loc(type.span) : fallbackLoc.value_or(currentLoc);
-  };
-  llvm::SmallVector<mlir::Type, 4> paramTypes;
-  for (const auto &param : fn.params) {
-    paramTypes.push_back(convertType(param.ty.value, typeLoc(param.ty)));
-  }
-
-  // Determine return type
-  llvm::SmallVector<mlir::Type, 1> resultTypes;
-  if (fn.return_type) {
-    auto retTy = convertType(fn.return_type->value, typeLoc(*fn.return_type));
-    // Don't add NoneType to results (unit return = no results)
-    if (!llvm::isa<mlir::NoneType>(retTy)) {
-      resultTypes.push_back(retTy);
-    }
-  }
-
-  // Prefer the per-function decl_span captured by the parser (points at the
-  // fn-name token) over the enclosing-item fallbackLoc.  This ensures impl
-  // methods get DW_AT_decl_line pointing at their own declaration line rather
-  // than the impl-block header line.
-  auto location = [&]() -> mlir::Location {
-    if (fn.decl_span && fn.decl_span->end > fn.decl_span->start)
-      return loc(*fn.decl_span);
-    return fallbackLoc.value_or(currentLoc);
-  }();
-
-  // Create the function at module scope.
-  // When nameOverride is provided, the caller has already mangled the name.
-  std::string funcName =
-      nameOverride.empty() ? mangleName(currentModulePath, "", fn.name) : nameOverride;
-
-  // fn main() without an explicit return type implicitly returns i32 exit
-  // code 0.  This keeps the POSIX ABI contract (int main) without forcing
-  // every program to write a trailing "0".  We defer adding i32 to
-  // resultTypes until after body generation so the body is generated in
-  // void context (no returnSlot/returnFlag, trailing expressions treated
-  // as statements).
-  bool isImplicitMainReturn = funcName == "main" && resultTypes.empty();
-
-  auto funcType = builder.getFunctionType(paramTypes, resultTypes);
-
-  // If a forward declaration exists (from pass 1.5), erase it — we'll
-  // replace it with the full definition that includes a body.
-  // If a full definition already exists (e.g. imported module's impl was
-  // already generated), skip — don't create a duplicate.
-  if (auto existing = module.lookupSymbol<mlir::func::FuncOp>(funcName)) {
-    if (existing.isDeclaration()) {
-      existing.erase();
-    } else {
-      // Already generated (e.g. from the defining module) — skip.
-      return existing;
-    }
-  }
-
-  auto savedIP = builder.saveInsertionPoint();
-  builder.setInsertionPointToEnd(module.getBody());
-  auto funcOp = mlir::func::FuncOp::create(builder, location, funcName, funcType);
-  setDebugParamNameAttrs(funcOp, fn.params, builder);
-
-  // Pub-aware linkage: main is always public, pub functions are public,
-  // everything else is private.
-  if (funcName == "main") {
-    // main is always public — no visibility change needed (default is public)
-  } else if (ast::is_pub(fn.visibility)) {
-    funcOp.setVisibility(mlir::SymbolTable::Visibility::Public);
-  } else {
-    funcOp.setVisibility(mlir::SymbolTable::Visibility::Private);
-  }
-
-  // Create entry block with parameter arguments
-  auto *entryBlock = funcOp.addEntryBlock();
-  builder.setInsertionPointToStart(entryBlock);
-
-  // Save and reset function-scoped state before parameter registration,
-  // so nested generateFunction calls (e.g., generics specialization) don't
-  // contaminate the outer function's allocas.
-  auto prevFunction = currentFunction;
-  currentFunction = funcOp;
-  auto prevFunctionReturnTypeExpr = currentFunctionReturnTypeExpr;
-  currentFunctionReturnTypeExpr = fn.return_type ? &fn.return_type->value : nullptr;
-
-  auto prevReturnFlag = returnFlag;
-  auto prevReturnSlot = returnSlot;
-  auto prevReturnSlotIsLazy = returnSlotIsLazy;
-  auto prevEarlyReturnFlag = earlyReturnFlag;
-  auto prevChannelIntOutValidAlloca = channelIntOutValidAlloca;
-  auto prevFuncLevelDropExcludeVars = std::move(funcLevelDropExcludeVars);
-  auto prevFuncLevelDropExcludeValues = std::move(funcLevelDropExcludeValues);
-  auto prevFuncLevelDropExcludeResolvedNames = std::move(funcLevelDropExcludeResolvedNames);
-  auto prevFuncLevelEarlyReturnExcludeValues = std::move(funcLevelEarlyReturnExcludeValues);
-  auto prevFuncLevelEarlyReturnExcludeResolvedNames =
-      std::move(funcLevelEarlyReturnExcludeResolvedNames);
-  auto prevFuncLevelDropScopeBase = funcLevelDropScopeBase;
-  auto prevPendingParamDrops = std::move(pendingFunctionParamDrops);
-  auto prevFnDefers = std::move(currentFnDefers);
-  returnFlag = nullptr;
-  returnSlot = nullptr;
-  returnSlotIsLazy = false;
-  earlyReturnFlag = nullptr;
-  channelIntOutValidAlloca = nullptr;
-  funcLevelDropExcludeVars.clear();
-  funcLevelDropExcludeValues.clear();
-  funcLevelDropExcludeResolvedNames.clear();
-  funcLevelEarlyReturnExcludeValues.clear();
-  funcLevelEarlyReturnExcludeResolvedNames.clear();
-  currentFnDefers.clear();
-  uint32_t paramIdx = 0;
-  for (const auto &param : fn.params) {
-    const auto &paramName = param.name;
-    auto paramValue = entryBlock->getArgument(paramIdx);
-    bindParam(param, paramValue);
-
-    // Track collection/handle/actor parameter types from type annotation
-    const auto &paramTy = param.ty.value;
-    {
-      auto resolveAliasExpr = [this](llvm::StringRef name) { return resolveTypeAliasExpr(name); };
-      auto handleStr = typeExprToHandleString(paramTy, knownHandleTypes, resolveAliasExpr);
-      if (!handleStr.empty()) {
-        handleVarTypes[paramName] = handleStr;
-        if (auto bindingIdentity = resolveCurrentBindingIdentity(paramName))
-          annotatedHandleTypes[bindingIdentity] = &paramTy;
-      }
-      if (auto streamInfo = streamHandleInfoFromTypeExpr(paramTy))
-        rememberTrackedStreamHandleInfo(paramName, *streamInfo);
-      auto actorName = typeExprToActorName(paramTy, resolveAliasExpr);
-      if (!actorName.empty() && actorRegistry.count(actorName))
-        actorVarTypes[paramName] = actorName;
-    }
-
-    // Track dyn Trait parameter types for dynamic dispatch
-    if (auto *traitObj = std::get_if<ast::TypeTraitObject>(&paramTy.kind)) {
-      if (!traitObj->bounds.empty()) {
-        dynTraitVarTypes[paramName] = traitObj->bounds[0].name;
-      }
-    }
-    ++paramIdx;
-  }
-
-  // Early return support: always create returnFlag so that `return` inside
-  // nested SCF regions (match arms, if branches) stores to the flag instead
-  // of emitting an illegal func.return inside an scf.if.
-  initReturnFlagAndSlot(resultTypes, location);
-
-  // Determine trailing expression variable names BEFORE generating the body.
-  // For struct init expressions, collect all field variable references to
-  // prevent double-free when the struct takes ownership of the values.
-  funcLevelDropExcludeVars.clear();
-  funcLevelDropScopeBase = dropScopes.size();
-  // Recursively collect identifiers from the return position of an
-  // expression (including if/match/block branches) so their owning
-  // variables are excluded from the function-level drop scope.
-  // Each candidate stores (name, depth) until lowering resolves it to a
-  // stable binding identity. The temporary depth tag lets us distinguish a
-  // returned outer binding from a shadowed inner binding with the same name.
-  // Determine whether the function body produces a value (non-void return).
-  // When producesValue is false, the last statement's result is discarded,
-  // so variables in it should NOT be excluded from drops.
-  bool fnProducesValue = false;
-  if (fn.return_type) {
-    // Check if return type is unit/void — if not, the body produces a value.
-    if (auto *named = std::get_if<ast::TypeNamed>(&fn.return_type->value.kind)) {
-      fnProducesValue = (named->name != "()" && named->name != "unit" && named->name != "void" &&
-                         named->name != "Never");
-    } else {
-      fnProducesValue = true; // Tuple, generic, etc. — non-void
-    }
-  }
-  collectExcludeVarsFromBlock(fn.body, funcLevelDropExcludeVars, 0, fnProducesValue);
-
-  // Build a flat (depth-independent) set of variable names appearing in the
-  // return expression.  Used by RAII close exclusion where the depth-based
-  // set doesn't work (enricher-wrapped unsafe blocks shift depths).
-  funcLevelReturnVarNames.clear();
-  for (const auto &[name, depth] : funcLevelDropExcludeVars)
-    funcLevelReturnVarNames.insert(name);
-  resolveFunctionDropExclusionCandidates();
-
-  // Build a separate set of vars referenced ONLY by explicit return
-  // statements. Used for path-specific drops when returnSlotIsLazy:
-  // the early-return path excludes these vars, the normal path does not.
-  funcLevelEarlyReturnVarNames.clear();
-  bool hasNestedReturn = false;
-  std::function<void(const ast::Block &)> scanReturns;
-  std::function<void(const ast::StmtIf &)> scanReturnsIf;
-  scanReturnsIf = [&scanReturns, &scanReturnsIf](const ast::StmtIf &ifStmt) {
-    scanReturns(ifStmt.then_block);
-    if (ifStmt.else_block) {
-      if (ifStmt.else_block->block)
-        scanReturns(*ifStmt.else_block->block);
-      if (ifStmt.else_block->if_stmt) {
-        auto &nested = ifStmt.else_block->if_stmt->value;
-        if (auto *nestedIf = std::get_if<ast::StmtIf>(&nested.kind))
-          scanReturnsIf(*nestedIf);
-      }
-    }
-  };
-  scanReturns = [&scanReturns, &scanReturnsIf, &hasNestedReturn, this](const ast::Block &blk) {
-    for (const auto &stmt : blk.stmts) {
-      if (auto *retStmt = std::get_if<ast::StmtReturn>(&stmt->value.kind)) {
-        hasNestedReturn = true;
-        if (retStmt->value) {
-          ExcludeSet tmp;
-          collectExcludeVars(retStmt->value->value, tmp, 0);
-          for (const auto &[name, d] : tmp)
-            funcLevelEarlyReturnVarNames.insert(name);
-        }
-      } else if (auto *ifStmt = std::get_if<ast::StmtIf>(&stmt->value.kind)) {
-        scanReturnsIf(*ifStmt);
-      } else if (auto *forStmt = std::get_if<ast::StmtFor>(&stmt->value.kind)) {
-        scanReturns(forStmt->body);
-      } else if (auto *whileStmt = std::get_if<ast::StmtWhile>(&stmt->value.kind)) {
-        scanReturns(whileStmt->body);
-      }
-    }
-  };
-  scanReturns(fn.body);
-
-  // Register param drops: use dropFuncForType (AST-based) to queue each param
-  // that carries ownership into pendingFunctionParamDrops.  generateBlock drains
-  // these into the body scope (depth 0) so the funcLevelDropExcludeVars pre-scan
-  // above already covers returned params and early-return paths correctly.
-  //
-  // Deliberately excluded:
-  //  • Non-handle-bearing __auto_field_drop structs (String / Vec / HashMap /
-  //    Closure-only wrappers): without alias-tracking we still cannot prove
-  //    that a returned field is not aliased in the result value.
-  //  • Handle types (Stream/Sink/Pair): these need closeAlloca RAII, not a
-  //    plain drop call.
-  // DROP-TODO(D3): BLOCKED — stream ownership convention is ambiguous per call
-  // site. A callee-side close-alloca would double-close handles that were
-  // consumed by downstream calls (e.g., hew_stream_pipe). Prerequisite: stream
-  // handle ownership model.
-  //  • Vec, HashMap, HashSet, Rc, bytes: these use borrow semantics at call
-  //    boundaries — the caller retains ownership and drops at scope exit.
-  //    Registering callee-side drops would cause double-frees whenever the
-  //    caller reuses the variable after the call (e.g. passing the same Vec
-  //    to multiple functions).
-  // DROP-TODO(D4): BLOCKED — no move-checker infrastructure exists.
-  // Borrow-semantics types (Vec, HashMap, HashSet, Rc, bytes) are reused after
-  // calls; enabling callee-side drops without single-use proof would cause
-  // use-after-free. Prerequisite: type-system move semantics.
-  //  • String: field aliases (e.g. `a.pre_release` passed to a fn) share the
-  //    heap buffer with the owning struct.  Callee-side drops would free that
-  //    buffer, and the owning struct's __auto_field_drop would free it again
-  //    → double-free.
-  // DROP-TODO(D5): BLOCKED — String params may alias struct fields (e.g.,
-  // a.pre_release shares buffer with struct a). No alias-tracking
-  // infrastructure exists to detect field aliasing. Prerequisite: same
-  // alias-tracking as D2.
-  //  • Indirect enum drops (`__hew_drop_*`): these functions are RECURSIVE —
-  //    they free the entire subtree.  When a pattern match passes children to
-  //    recursive callee calls that drop them, the parent's param drop would
-  //    then drop them again → double-free.
-  //  • Self-referential drop functions: a user Drop's own parameter must not
-  //    be re-dropped — that would recurse into the Drop impl infinitely.
-  auto isBorrowSemanticsDrop = [](const std::string &df) {
-    return df == "hew_vec_free" || df == "hew_hashmap_free_impl" || df == "hew_hashset_free" ||
-           df == "hew_rc_drop" || df == "hew_string_drop" ||
-           df.starts_with("__hew_drop_"); // indirect enum recursive drops
-  };
-  for (const auto &param : fn.params) {
-    if (handleVarTypes.count(param.name) || lookupTrackedStreamHandleInfo(param.name))
-      continue;
-    auto dropFunc = dropFuncForType(param.ty.value);
-    if (dropFunc.empty() || isBorrowSemanticsDrop(dropFunc))
-      continue;
-    if (dropFunc == "__auto_field_drop") {
-      auto *named = std::get_if<ast::TypeNamed>(&param.ty.value.kind);
-      if (!named || !isHandleBearingStruct(resolveTypeAlias(named->name)))
-        continue;
-    }
-    // Skip if this function IS the drop function for the param's type.
-    // A user Drop implementation must not re-drop its own parameter — doing so
-    // would recurse into itself infinitely.
-    if (dropFunc == funcName)
-      continue;
-    bool isUserDrop = false;
-    if (auto *named = std::get_if<ast::TypeNamed>(&param.ty.value.kind)) {
-      auto typeName = resolveTypeAlias(named->name);
-      isUserDrop = userDropFuncs.count(typeName) > 0;
-    }
-    pendingFunctionParamDrops.push_back({param.name, dropFunc, isUserDrop});
-  }
-
-  // Fail-closed: reject returning a field of a callee-dropped parameter.
-  // Without full field-alias / partial-move tracking, returning `param.field`
-  // when `param` gets a callee-side drop would either double-free the field
-  // (if the param IS dropped) or leak sibling fields (if we skip the drop).
-  // Reject this shape at compile time until alias-tracking lands.
-  if (!pendingFunctionParamDrops.empty() && fnProducesValue) {
-    std::unordered_set<std::string> droppedParamNames;
-    for (const auto &pd : pendingFunctionParamDrops)
-      droppedParamNames.insert(pd.name);
-
-    // Recursively check whether an expression is (or wraps) a field access
-    // on a callee-dropped param.  Recurses through block/if/if-let/match/
-    // unsafe/scope wrappers so that wrapped aliases are fail-closed.
-    //
-    // blockValueYieldsFieldOfDroppedParam checks the value-position of a
-    // Block: trailing expr first, then (when absent) the last statement's
-    // expression / StmtIf / StmtMatch / loop-family (via break values).
-    // This mirrors collectExcludeVarsFromBlock.
-    //
-    // blockBreakValueYieldsFieldOfDroppedParam scans a loop body for
-    // StmtBreak nodes whose break-value is a field of a dropped param,
-    // recursing through nested control flow (StmtIf/else-if chains,
-    // StmtIfLet, StmtMatch, StmtExpression wrappers, and nested loops for
-    // labeled breaks).
-    //
-    // SCOPE: defense-in-depth — the type checker currently rejects loop
-    // break values (loops produce ()) so the codegen guard doesn't fire
-    // today.  This covers the case where typed break-values are implemented.
-    std::function<bool(const ast::Expr &)> isFieldOfDroppedParam;
-    std::function<bool(const ast::Block &)> blockValueYieldsFieldOfDroppedParam;
-    std::function<bool(const ast::StmtIf &)> stmtIfValueYieldsFieldOfDroppedParam;
-    std::function<bool(const ast::Block &, const std::optional<std::string> &, bool)>
-        blockBreakValueYieldsFieldOfDroppedParam;
-    std::function<bool(const ast::Expr &, const std::optional<std::string> &, bool)>
-        exprBreakValueYieldsFieldOfDroppedParam;
-    std::function<bool(const ast::StmtIf &, const std::optional<std::string> &, bool)>
-        stmtIfBreakValueYieldsFieldOfDroppedParam;
-    stmtIfValueYieldsFieldOfDroppedParam = [&](const ast::StmtIf &ifStmt) -> bool {
-      if (blockValueYieldsFieldOfDroppedParam(ifStmt.then_block))
-        return true;
-      if (ifStmt.else_block) {
-        if (ifStmt.else_block->block &&
-            blockValueYieldsFieldOfDroppedParam(*ifStmt.else_block->block))
-          return true;
-        if (ifStmt.else_block->if_stmt) {
-          if (auto *nested = std::get_if<ast::StmtIf>(&ifStmt.else_block->if_stmt->value.kind))
-            if (stmtIfValueYieldsFieldOfDroppedParam(*nested))
-              return true;
-          // else if let: check both branches for value-position field access.
-          if (auto *nestedIfLet =
-                  std::get_if<ast::StmtIfLet>(&ifStmt.else_block->if_stmt->value.kind)) {
-            if (blockValueYieldsFieldOfDroppedParam(nestedIfLet->body))
-              return true;
-            if (nestedIfLet->else_body &&
-                blockValueYieldsFieldOfDroppedParam(*nestedIfLet->else_body))
-              return true;
-          }
-        }
-      }
-      return false;
-    };
-    // Scan a StmtIf (and else-if chains) inside a loop body for break-values.
-    stmtIfBreakValueYieldsFieldOfDroppedParam = [&](const ast::StmtIf &ifStmt,
-                                                    const std::optional<std::string> &targetLabel,
-                                                    bool checkUnlabeled) -> bool {
-      if (blockBreakValueYieldsFieldOfDroppedParam(ifStmt.then_block, targetLabel, checkUnlabeled))
-        return true;
-      if (ifStmt.else_block) {
-        if (ifStmt.else_block->block && blockBreakValueYieldsFieldOfDroppedParam(
-                                            *ifStmt.else_block->block, targetLabel, checkUnlabeled))
-          return true;
-        if (ifStmt.else_block->if_stmt) {
-          if (auto *nested = std::get_if<ast::StmtIf>(&ifStmt.else_block->if_stmt->value.kind))
-            if (stmtIfBreakValueYieldsFieldOfDroppedParam(*nested, targetLabel, checkUnlabeled))
-              return true;
-          if (auto *nestedIfLet =
-                  std::get_if<ast::StmtIfLet>(&ifStmt.else_block->if_stmt->value.kind)) {
-            if (blockBreakValueYieldsFieldOfDroppedParam(nestedIfLet->body, targetLabel,
-                                                         checkUnlabeled))
-              return true;
-            if (nestedIfLet->else_body && blockBreakValueYieldsFieldOfDroppedParam(
-                                              *nestedIfLet->else_body, targetLabel, checkUnlabeled))
-              return true;
-          }
-        }
-      }
-      return false;
-    };
-    // Scan an expression for break-values inside a loop body.
-    exprBreakValueYieldsFieldOfDroppedParam = [&](const ast::Expr &expr,
-                                                  const std::optional<std::string> &targetLabel,
-                                                  bool checkUnlabeled) -> bool {
-      if (auto *blockE = std::get_if<ast::ExprBlock>(&expr.kind))
-        return blockBreakValueYieldsFieldOfDroppedParam(blockE->block, targetLabel, checkUnlabeled);
-      if (auto *unsafeE = std::get_if<ast::ExprUnsafe>(&expr.kind))
-        return blockBreakValueYieldsFieldOfDroppedParam(unsafeE->block, targetLabel,
-                                                        checkUnlabeled);
-      if (auto *scopeE = std::get_if<ast::ExprScope>(&expr.kind))
-        return blockBreakValueYieldsFieldOfDroppedParam(scopeE->block, targetLabel, checkUnlabeled);
-      if (auto *ifE = std::get_if<ast::ExprIf>(&expr.kind)) {
-        if (ifE->then_block && exprBreakValueYieldsFieldOfDroppedParam(ifE->then_block->value,
-                                                                       targetLabel, checkUnlabeled))
-          return true;
-        if (ifE->else_block && *ifE->else_block &&
-            exprBreakValueYieldsFieldOfDroppedParam((*ifE->else_block)->value, targetLabel,
-                                                    checkUnlabeled))
-          return true;
-        return false;
-      }
-      if (auto *ifLet = std::get_if<ast::ExprIfLet>(&expr.kind)) {
-        if (blockBreakValueYieldsFieldOfDroppedParam(ifLet->body, targetLabel, checkUnlabeled))
-          return true;
-        if (ifLet->else_body && blockBreakValueYieldsFieldOfDroppedParam(
-                                    *ifLet->else_body, targetLabel, checkUnlabeled))
-          return true;
-        return false;
-      }
-      if (auto *matchE = std::get_if<ast::ExprMatch>(&expr.kind)) {
-        for (const auto &arm : matchE->arms)
-          if (arm.body &&
-              exprBreakValueYieldsFieldOfDroppedParam(arm.body->value, targetLabel, checkUnlabeled))
-            return true;
-        return false;
-      }
-      return false;
-    };
-    // Scan a loop body block for StmtBreak nodes whose break-value is a field
-    // of a dropped param.  See blockBreakYieldsFieldAccess in MLIRGenExpr.cpp
-    // for detailed parameter semantics (targetLabel, checkUnlabeled).
-    blockBreakValueYieldsFieldOfDroppedParam = [&](const ast::Block &blk,
-                                                   const std::optional<std::string> &targetLabel,
-                                                   bool checkUnlabeled) -> bool {
-      for (const auto &stmt : blk.stmts) {
-        if (auto *brk = std::get_if<ast::StmtBreak>(&stmt->value.kind)) {
-          if (brk->value) {
-            bool targets_us = false;
-            if (!brk->label && checkUnlabeled)
-              targets_us = true;
-            if (brk->label && targetLabel && *brk->label == *targetLabel)
-              targets_us = true;
-            if (targets_us && isFieldOfDroppedParam(brk->value->value))
-              return true;
-          }
-          continue;
-        }
-        if (auto *ifStmt = std::get_if<ast::StmtIf>(&stmt->value.kind)) {
-          if (stmtIfBreakValueYieldsFieldOfDroppedParam(*ifStmt, targetLabel, checkUnlabeled))
-            return true;
-          continue;
-        }
-        if (auto *ifLetStmt = std::get_if<ast::StmtIfLet>(&stmt->value.kind)) {
-          if (blockBreakValueYieldsFieldOfDroppedParam(ifLetStmt->body, targetLabel,
-                                                       checkUnlabeled))
-            return true;
-          if (ifLetStmt->else_body && blockBreakValueYieldsFieldOfDroppedParam(
-                                          *ifLetStmt->else_body, targetLabel, checkUnlabeled))
-            return true;
-          continue;
-        }
-        // StmtMatch arm bodies are statement contexts inside a loop —
-        // scan them for breaks, not as result expressions.
-        if (auto *matchStmt = std::get_if<ast::StmtMatch>(&stmt->value.kind)) {
-          for (const auto &arm : matchStmt->arms)
-            if (arm.body && exprBreakValueYieldsFieldOfDroppedParam(arm.body->value, targetLabel,
-                                                                    checkUnlabeled))
-              return true;
-          continue;
-        }
-        if (auto *exprStmt = std::get_if<ast::StmtExpression>(&stmt->value.kind)) {
-          if (exprBreakValueYieldsFieldOfDroppedParam(exprStmt->expr.value, targetLabel,
-                                                      checkUnlabeled))
-            return true;
-          continue;
-        }
-        // Nested loops: only labeled breaks targeting our loop escape.
-        // If the inner loop REUSES the same label name (shadowing), breaks
-        // with that label inside target the inner loop — pass nullopt so
-        // they are no longer attributed to the outer loop.
-        if (auto *innerLoop = std::get_if<ast::StmtLoop>(&stmt->value.kind)) {
-          const auto innerTarget =
-              (innerLoop->label && targetLabel && *innerLoop->label == *targetLabel)
-                  ? std::optional<std::string>{}
-                  : targetLabel;
-          if (blockBreakValueYieldsFieldOfDroppedParam(innerLoop->body, innerTarget, false))
-            return true;
-          continue;
-        }
-        if (auto *innerWhile = std::get_if<ast::StmtWhile>(&stmt->value.kind)) {
-          const auto innerTarget =
-              (innerWhile->label && targetLabel && *innerWhile->label == *targetLabel)
-                  ? std::optional<std::string>{}
-                  : targetLabel;
-          if (blockBreakValueYieldsFieldOfDroppedParam(innerWhile->body, innerTarget, false))
-            return true;
-          continue;
-        }
-        if (auto *innerWhileLet = std::get_if<ast::StmtWhileLet>(&stmt->value.kind)) {
-          const auto innerTarget =
-              (innerWhileLet->label && targetLabel && *innerWhileLet->label == *targetLabel)
-                  ? std::optional<std::string>{}
-                  : targetLabel;
-          if (blockBreakValueYieldsFieldOfDroppedParam(innerWhileLet->body, innerTarget, false))
-            return true;
-          continue;
-        }
-        if (auto *innerFor = std::get_if<ast::StmtFor>(&stmt->value.kind)) {
-          const auto innerTarget =
-              (innerFor->label && targetLabel && *innerFor->label == *targetLabel)
-                  ? std::optional<std::string>{}
-                  : targetLabel;
-          if (blockBreakValueYieldsFieldOfDroppedParam(innerFor->body, innerTarget, false))
-            return true;
-          continue;
-        }
-      }
-      return false;
-    };
-    blockValueYieldsFieldOfDroppedParam = [&](const ast::Block &blk) -> bool {
-      if (blk.trailing_expr)
-        return isFieldOfDroppedParam(blk.trailing_expr->value);
-      if (!blk.stmts.empty()) {
-        const auto &last = blk.stmts.back()->value;
-        if (auto *exprStmt = std::get_if<ast::StmtExpression>(&last.kind))
-          return isFieldOfDroppedParam(exprStmt->expr.value);
-        if (auto *ifStmt = std::get_if<ast::StmtIf>(&last.kind))
-          return stmtIfValueYieldsFieldOfDroppedParam(*ifStmt);
-        if (auto *ifLetStmt = std::get_if<ast::StmtIfLet>(&last.kind)) {
-          if (blockValueYieldsFieldOfDroppedParam(ifLetStmt->body))
-            return true;
-          if (ifLetStmt->else_body && blockValueYieldsFieldOfDroppedParam(*ifLetStmt->else_body))
-            return true;
-        }
-        if (auto *matchStmt = std::get_if<ast::StmtMatch>(&last.kind)) {
-          for (const auto &arm : matchStmt->arms)
-            if (arm.body && isFieldOfDroppedParam(arm.body->value))
-              return true;
-        }
-        if (auto *loopStmt = std::get_if<ast::StmtLoop>(&last.kind))
-          return blockBreakValueYieldsFieldOfDroppedParam(loopStmt->body, loopStmt->label, true);
-        if (auto *whileStmt = std::get_if<ast::StmtWhile>(&last.kind))
-          return blockBreakValueYieldsFieldOfDroppedParam(whileStmt->body, whileStmt->label, true);
-        if (auto *whileLetStmt = std::get_if<ast::StmtWhileLet>(&last.kind))
-          return blockBreakValueYieldsFieldOfDroppedParam(whileLetStmt->body, whileLetStmt->label,
-                                                          true);
-        if (auto *forStmt = std::get_if<ast::StmtFor>(&last.kind))
-          return blockBreakValueYieldsFieldOfDroppedParam(forStmt->body, forStmt->label, true);
-      }
-      return false;
-    };
-    isFieldOfDroppedParam = [&](const ast::Expr &expr) -> bool {
-      return exprYieldsFieldMatching(expr, [&](const ast::ExprFieldAccess &fieldAccess) {
-        auto *id = std::get_if<ast::ExprIdentifier>(&fieldAccess.object->value.kind);
-        return id && droppedParamNames.count(id->name);
-      });
-    };
-
-    // Scan value-position of the function body (trailing expr or last stmt)
-    if (blockValueYieldsFieldOfDroppedParam(fn.body)) {
-      ++errorCount_;
-      emitError(location) << "returning a field of an owned parameter is not yet supported "
-                          << "(field-alias ownership tracking is required); "
-                          << "return the whole parameter or clone the field instead";
-      return nullptr;
-    }
-
-    // Scan explicit return statements (recursively)
-    std::function<bool(const ast::Block &)> scanForFieldReturn;
-    std::function<bool(const ast::StmtIf &)> scanForFieldReturnIf;
-    scanForFieldReturnIf = [&](const ast::StmtIf &ifStmt) -> bool {
-      if (scanForFieldReturn(ifStmt.then_block))
-        return true;
-      if (ifStmt.else_block) {
-        if (ifStmt.else_block->block && scanForFieldReturn(*ifStmt.else_block->block))
-          return true;
-        if (ifStmt.else_block->if_stmt) {
-          if (auto *nested = std::get_if<ast::StmtIf>(&ifStmt.else_block->if_stmt->value.kind))
-            if (scanForFieldReturnIf(*nested))
-              return true;
-        }
-      }
-      return false;
-    };
-    scanForFieldReturn = [&](const ast::Block &blk) -> bool {
-      for (const auto &stmt : blk.stmts) {
-        if (auto *ret = std::get_if<ast::StmtReturn>(&stmt->value.kind)) {
-          if (ret->value && isFieldOfDroppedParam(ret->value->value)) {
-            ++errorCount_;
-            emitError(location) << "returning a field of an owned parameter is not yet supported "
-                                << "(field-alias ownership tracking is required); "
-                                << "return the whole parameter or clone the field instead";
-            return true;
-          }
-        } else if (auto *ifStmt = std::get_if<ast::StmtIf>(&stmt->value.kind)) {
-          if (scanForFieldReturnIf(*ifStmt))
-            return true;
-        } else if (auto *forStmt = std::get_if<ast::StmtFor>(&stmt->value.kind)) {
-          if (scanForFieldReturn(forStmt->body))
-            return true;
-        } else if (auto *whileStmt = std::get_if<ast::StmtWhile>(&stmt->value.kind)) {
-          if (scanForFieldReturn(whileStmt->body))
-            return true;
-        }
-      }
-      return false;
-    };
-    if (scanForFieldReturn(fn.body))
-      return nullptr;
-  }
-
-  // If the pre-scan found return statements in nested scopes (if/for/while)
-  // and the function returns an aggregate type that didn't get an eager
-  // returnSlot, create the slot now so useReturnGuards activates in
-  // generateBlock. Without this, statements after the early return would
-  // execute unconditionally (creating and leaking temporaries).
-  if (hasNestedReturn && returnFlag && !returnSlot) {
-    ensureReturnSlot(location);
-  }
-
-  // Generate the function body. Functions with an explicit unit/void return
-  // type (or implicit main) discard the block result. For unannotated functions
-  // that already contain nested explicit returns, keep final statement-position
-  // if/match lowering on the statement path rather than forcing a value path
-  // with no established function result type.
-  bool finalStmtNeedsStatementLowering = false;
-  if (!fn.return_type && hasNestedReturn && !fn.body.trailing_expr && !fn.body.stmts.empty()) {
-    const auto &lastStmt = fn.body.stmts.back()->value;
-    finalStmtNeedsStatementLowering = std::holds_alternative<ast::StmtIf>(lastStmt.kind) ||
-                                      std::holds_alternative<ast::StmtMatch>(lastStmt.kind);
-  }
-  bool bodyResultDiscarded = isImplicitMainReturn || (fn.return_type && resultTypes.empty()) ||
-                             finalStmtNeedsStatementLowering;
-  mlir::Value bodyValue = generateBlock(fn.body, /*statementPosition=*/bodyResultDiscarded,
-                                        /*isFunctionBodyBlock=*/true);
-  funcLevelDropExcludeVars.clear();
-  funcLevelReturnVarNames.clear();
-  funcLevelEarlyReturnVarNames.clear();
-
-  // Infer return type from body if not explicitly annotated (skip for
-  // implicit main — we handle its return separately below).
-  if (!isImplicitMainReturn && resultTypes.empty() && bodyValue && bodyValue.getType() &&
-      !llvm::isa<mlir::NoneType>(bodyValue.getType())) {
-    resultTypes.push_back(bodyValue.getType());
-    auto newFuncType = builder.getFunctionType(paramTypes, resultTypes);
-    funcOp.setFunctionType(newFuncType);
-  }
-
-  // Handle return: check whether early return was taken or use body value
-  auto *currentBlock = builder.getInsertionBlock();
-  if (currentBlock &&
-      (currentBlock->empty() || !currentBlock->back().hasTrait<mlir::OpTrait::IsTerminator>())) {
-    if (returnFlag && returnSlot && !resultTypes.empty()) {
-      // Select between returnSlot (early return) and bodyValue (normal flow).
-      // Drops were already emitted path-specifically in popDropScope.
-      auto flagVal =
-          mlir::memref::LoadOp::create(builder, location, returnFlag, mlir::ValueRange{});
-
-      auto selectOp = mlir::scf::IfOp::create(builder, location, resultTypes[0], flagVal,
-                                              /*withElseRegion=*/true);
-
-      // Then (early return was taken): yield slot value
-      builder.setInsertionPointToStart(&selectOp.getThenRegion().front());
-      auto slotVal = mlir::memref::LoadOp::create(builder, location, returnSlot, mlir::ValueRange{})
-                         .getResult();
-      slotVal = coerceTypeForSink(slotVal, resultTypes[0], location);
-      mlir::scf::YieldOp::create(builder, location, mlir::ValueRange{slotVal});
-
-      // Else (normal flow): yield body value
-      builder.setInsertionPointToStart(&selectOp.getElseRegion().front());
-      mlir::Value normalValue = bodyValue;
-      if (!normalValue) {
-        normalValue = createDefaultValue(builder, location, resultTypes[0]);
-      }
-      normalValue = coerceTypeForSink(normalValue, resultTypes[0], location);
-      mlir::scf::YieldOp::create(builder, location, mlir::ValueRange{normalValue});
-
-      builder.setInsertionPointAfter(selectOp);
-      mlir::func::ReturnOp::create(builder, location, mlir::ValueRange{selectOp.getResult(0)});
-    } else if (bodyValue && !resultTypes.empty()) {
-      // Emit drops before return, excluding the exact bindings whose
-      // ownership was transferred into the function result.
-      if (!funcLevelDropExcludeValues.empty())
-        emitDropsExcept(funcLevelDropExcludeValues);
-      else
-        emitAllDrops();
-      auto coercedBody = coerceTypeForSink(bodyValue, resultTypes[0], location);
-      mlir::func::ReturnOp::create(builder, location, mlir::ValueRange{coercedBody});
-    } else {
-      emitAllDrops();
-      if (isImplicitMainReturn) {
-        // Patch the function type to return i32 and emit return 0
-        resultTypes.push_back(builder.getI32Type());
-        funcOp.setFunctionType(builder.getFunctionType(paramTypes, resultTypes));
-        auto zero = createIntConstant(builder, location, builder.getI32Type(), 0);
-        mlir::func::ReturnOp::create(builder, location, mlir::ValueRange{zero});
-      } else {
-        mlir::func::ReturnOp::create(builder, location);
-      }
-    }
-  }
-
-  // Restore previous function state
-  returnFlag = prevReturnFlag;
-  returnSlot = prevReturnSlot;
-  returnSlotIsLazy = prevReturnSlotIsLazy;
-  earlyReturnFlag = prevEarlyReturnFlag;
-  channelIntOutValidAlloca = prevChannelIntOutValidAlloca;
-  funcLevelDropExcludeVars = std::move(prevFuncLevelDropExcludeVars);
-  funcLevelDropExcludeValues = std::move(prevFuncLevelDropExcludeValues);
-  funcLevelDropExcludeResolvedNames = std::move(prevFuncLevelDropExcludeResolvedNames);
-  funcLevelEarlyReturnExcludeValues = std::move(prevFuncLevelEarlyReturnExcludeValues);
-  funcLevelEarlyReturnExcludeResolvedNames =
-      std::move(prevFuncLevelEarlyReturnExcludeResolvedNames);
-  funcLevelDropScopeBase = prevFuncLevelDropScopeBase;
-  pendingFunctionParamDrops = std::move(prevPendingParamDrops);
-  currentFnDefers = std::move(prevFnDefers);
-  currentFunctionReturnTypeExpr = prevFunctionReturnTypeExpr;
-  currentFunction = prevFunction;
-  builder.restoreInsertionPoint(savedIP);
-  return funcOp;
-}
-
-// ============================================================================
 // Generator function codegen (LLVM coroutine-based)
 // ============================================================================
 //
