+++ hew-codegen/src/mlir/MLIRGenExpr.cpp
@@ -55,816 +55,6 @@ static bool isOwnershipExemptDropFunc(llvm::StringRef dropFunc) {
 }
 
 // ============================================================================
-// Expression generation
-// ============================================================================
-
-mlir::Value MLIRGen::generateExpression(const ast::Expr &expr, std::optional<mlir::Type> typeHint) {
-  currentLoc = loc(expr.span);
-
-  // Return pre-computed value for hoisted loop-invariant sub-expressions.
-  auto hoistIt = hoistedValues.find(&expr);
-  if (hoistIt != hoistedValues.end())
-    return hoistIt->second;
-
-  if (auto *lit = std::get_if<ast::ExprLiteral>(&expr.kind))
-    return generateLiteral(lit->lit, expr.span);
-
-  if (auto *ident = std::get_if<ast::ExprIdentifier>(&expr.kind)) {
-    auto name = ident->name;
-    // Check regular variables first
-    auto val = lookupVariable(name);
-    if (val)
-      return val;
-
-    // Inside actor init/receive, bare field names (e.g. `handle`) resolve to
-    // self.field via GEP into the actor state struct.
-    if (!currentActorName.empty()) {
-      auto selfVal = lookupVariable("self");
-      if (selfVal) {
-        auto actorIt = structTypes.find(currentActorName);
-        if (actorIt != structTypes.end()) {
-          auto structType = mlir::dyn_cast<mlir::LLVM::LLVMStructType>(actorIt->second.mlirType);
-          if (structType) {
-            for (const auto &field : actorIt->second.fields) {
-              if (field.name == name) {
-                auto ptrType = mlir::LLVM::LLVMPointerType::get(&context);
-                auto fieldPtr = mlir::LLVM::GEPOp::create(
-                    builder, currentLoc, ptrType, structType, selfVal,
-                    llvm::ArrayRef<mlir::LLVM::GEPArg>{0, static_cast<int32_t>(field.index)});
-                auto fieldVal =
-                    mlir::LLVM::LoadOp::create(builder, currentLoc, field.type, fieldPtr)
-                        .getResult();
-                if ((mlir::isa<hew::VecType>(field.semanticType) ||
-                     mlir::isa<hew::HashMapType>(field.semanticType)) &&
-                    field.semanticType != field.type)
-                  return coerceType(fieldVal, field.semanticType, currentLoc);
-                return fieldVal;
-              }
-            }
-          }
-        }
-      }
-    }
-
-    // Check module-level constants
-    auto constIt = moduleConstants.find(name);
-    if (constIt != moduleConstants.end()) {
-      return generateExpression(*constIt->second);
-    }
-
-    // Check enum unit variants (e.g., Red, Green, Blue, None)
-    auto varIt = variantLookup.find(name);
-    if (varIt != variantLookup.end()) {
-      const auto &enumName = varIt->second.first;
-      auto variantIndex = static_cast<int64_t>(varIt->second.second);
-
-      // Built-in None: construct Option { tag=0 }
-      if (name == "None" && enumName == "__Option") {
-        auto location = currentLoc;
-        mlir::Type optionType = resolveOptionConstructorType(typeHint, expr.span);
-        if (!optionType) {
-          ++errorCount_;
-          emitError(location)
-              << "cannot determine type for `None`; add an explicit type annotation";
-          return nullptr;
-        }
-        // No payload → mlir::ArrayAttr{} (null) is intentionally absent.
-        // EnumConstructOp::verify() skips position checking when payloads
-        // is empty, so the null ArrayAttr is semantically correct here.
-        mlir::Value result = hew::EnumConstructOp::create(
-            builder, location, optionType, static_cast<uint32_t>(variantIndex),
-            llvm::StringRef("Option"), mlir::ValueRange{},
-            /*payload_positions=*/mlir::ArrayAttr{});
-        return result;
-      }
-
-      auto enumIt = enumTypes.find(enumName);
-      if (enumIt != enumTypes.end()) {
-        const auto &enumInfo = enumIt->second;
-
-        if (enumInfo.hasPayloads) {
-          // Unit variant of a payload enum: build struct { tag, undef... }
-          auto location = currentLoc;
-
-          if (enumInfo.isIndirect) {
-            // Indirect enum: allocate on heap
-            auto innerType = enumInfo.innerStructType;
-            mlir::Value structVal = hew::EnumConstructOp::create(
-                builder, location, innerType, static_cast<uint32_t>(variantIndex),
-                llvm::StringRef(enumName), mlir::ValueRange{},
-                /*payload_positions=*/mlir::ArrayAttr{});
-            auto ptrType = mlir::LLVM::LLVMPointerType::get(&context);
-            auto mallocFuncType = mlir::FunctionType::get(&context, {sizeType()}, {ptrType});
-            getOrCreateExternFunc("malloc", mallocFuncType);
-            auto sizeVal = hew::SizeOfOp::create(builder, location, sizeType(),
-                                                 mlir::TypeAttr::get(innerType));
-            auto mallocCall = mlir::func::CallOp::create(
-                builder, location, "malloc", mlir::TypeRange{ptrType}, mlir::ValueRange{sizeVal});
-            auto allocPtr = mallocCall.getResult(0);
-            mlir::LLVM::StoreOp::create(builder, location, structVal, allocPtr);
-            return allocPtr;
-          }
-
-          mlir::Value result = hew::EnumConstructOp::create(
-              builder, location, enumInfo.mlirType, static_cast<uint32_t>(variantIndex),
-              llvm::StringRef(enumName), mlir::ValueRange{},
-              /*payload_positions=*/mlir::ArrayAttr{});
-          return result;
-        }
-        // All-unit enum: just produce the tag index as i32
-        return createIntConstant(builder, currentLoc, builder.getI32Type(), variantIndex);
-      }
-
-      // Built-in variant with no enumTypes entry
-      return createIntConstant(builder, currentLoc, builder.getI32Type(), variantIndex);
-    }
-
-    // Check if the name refers to a module-level function (function-as-value)
-    std::string mangledFuncName = mangleName(currentModulePath, "", name);
-    auto funcOp = module.lookupSymbol<mlir::func::FuncOp>(mangledFuncName);
-    if (!funcOp)
-      funcOp = module.lookupSymbol<mlir::func::FuncOp>(name); // fallback to unmangled
-    if (funcOp) {
-      auto symName = funcOp.getSymName();
-      return mlir::func::ConstantOp::create(builder, currentLoc, funcOp.getFunctionType(),
-                                            mlir::SymbolRefAttr::get(&context, symName));
-    }
-
-    ++errorCount_;
-    emitError(currentLoc) << "undeclared variable '" << name
-                          << "'; did you mean to declare it with 'let' or 'var'?";
-    return nullptr;
-  }
-
-  if (auto *bin = std::get_if<ast::ExprBinary>(&expr.kind))
-    return generateBinaryExpr(*bin);
-  if (auto *un = std::get_if<ast::ExprUnary>(&expr.kind))
-    return generateUnaryExpr(*un);
-  if (auto *call = std::get_if<ast::ExprCall>(&expr.kind))
-    return generateCallExpr(*call, expr.span, typeHint);
-  if (auto *ifE = std::get_if<ast::ExprIf>(&expr.kind))
-    return generateIfExpr(*ifE, expr.span);
-  if (auto *blockExpr = std::get_if<ast::ExprBlock>(&expr.kind)) {
-    // Empty block {} coerces to HashMap when the surrounding expression carries
-    // an explicit collection hint.
-    if (blockExpr->block.stmts.empty() && !blockExpr->block.trailing_expr && typeHint &&
-        mlir::isa<hew::HashMapType>(*typeHint)) {
-      return hew::HashMapNewOp::create(builder, currentLoc, *typeHint).getResult();
-    }
-    auto blockResult = generateBlockExpr(blockExpr->block);
-    if (!blockResult)
-      return nullptr;
-    return blockResult;
-  }
-  if (auto *cast = std::get_if<ast::ExprCast>(&expr.kind)) {
-    auto location = currentLoc;
-    auto value = generateExpression(cast->expr->value);
-    if (!value)
-      return nullptr;
-    currentLoc = location;
-    auto targetType = convertTypeOrError(cast->ty.value, "cannot resolve cast target type");
-    if (!targetType)
-      return nullptr;
-    bool isUnsigned = isUnsignedTypeExpr(cast->ty.value);
-    return coerceType(value, targetType, location, isUnsigned);
-  }
-  if (auto *pf = std::get_if<ast::ExprPostfixTry>(&expr.kind))
-    return generatePostfixExpr(*pf);
-  if (auto *me = std::get_if<ast::ExprMatch>(&expr.kind))
-    return generateMatchExpr(*me, expr.span);
-  if (auto *se = std::get_if<ast::ExprScope>(&expr.kind))
-    return generateScopeExpr(*se);
-  if (auto *sle = std::get_if<ast::ExprScopeLaunch>(&expr.kind))
-    return generateScopeLaunchExpr(*sle);
-  if (auto *sse = std::get_if<ast::ExprScopeSpawn>(&expr.kind))
-    return generateScopeSpawnExpr(*sse);
-  if (std::get_if<ast::ExprScopeCancel>(&expr.kind))
-    return generateScopeCancelExpr();
-
-  if (auto *sel = std::get_if<ast::ExprSelect>(&expr.kind))
-    return generateSelectExpr(*sel);
-  if (auto *join = std::get_if<ast::ExprJoin>(&expr.kind))
-    return generateJoinExpr(*join);
-  if (auto *spawn = std::get_if<ast::ExprSpawn>(&expr.kind))
-    return generateSpawnExpr(*spawn);
-  if (auto *spawnLambda = std::get_if<ast::ExprSpawnLambdaActor>(&expr.kind))
-    return generateSpawnLambdaActorExpr(*spawnLambda);
-  if (auto *send = std::get_if<ast::ExprSend>(&expr.kind))
-    return generateSendExpr(*send);
-  if (auto *si = std::get_if<ast::ExprStructInit>(&expr.kind))
-    return generateStructInit(*si, expr.span);
-  if (auto *mc = std::get_if<ast::ExprMethodCall>(&expr.kind))
-    return generateMethodCall(*mc, expr.span);
-  if (auto *tup = std::get_if<ast::ExprTuple>(&expr.kind))
-    return generateTupleExpr(*tup, typeHint);
-  if (auto *arr = std::get_if<ast::ExprArray>(&expr.kind))
-    return generateArrayExpr(*arr, typeHint);
-  if (auto *mapLit = std::get_if<ast::ExprMapLiteral>(&expr.kind))
-    return generateMapLiteralExpr(*mapLit, expr.span, typeHint);
-  if (auto *lam = std::get_if<ast::ExprLambda>(&expr.kind))
-    return generateLambdaExpr(*lam);
-  if (auto *interp = std::get_if<ast::ExprInterpolatedString>(&expr.kind))
-    return generateInterpolatedString(*interp);
-  if (auto *regex = std::get_if<ast::ExprRegexLiteral>(&expr.kind))
-    return generateRegexLiteral(*regex);
-  if (auto *bsl = std::get_if<ast::ExprByteStringLiteral>(&expr.kind))
-    return generateBytesLiteral(bsl->data);
-  if (auto *bal = std::get_if<ast::ExprByteArrayLiteral>(&expr.kind))
-    return generateBytesLiteral(bal->data);
-  if (auto *ifLet = std::get_if<ast::ExprIfLet>(&expr.kind))
-    return generateIfLetExpr(*ifLet, expr.span);
-  if (auto *arrRepeat = std::get_if<ast::ExprArrayRepeat>(&expr.kind))
-    return generateArrayRepeatExpr(*arrRepeat, expr.span);
-  if (auto *to = std::get_if<ast::ExprTimeout>(&expr.kind)) {
-    // Evaluate duration for validation but discard; timeout not yet enforced
-    generateExpression(to->duration->value);
-    return generateExpression(to->expr->value);
-  }
-
-  if (auto *ue = std::get_if<ast::ExprUnsafe>(&expr.kind)) {
-    auto unsafeResult = generateBlock(ue->block);
-    if (!unsafeResult)
-      return nullptr;
-    return unsafeResult;
-  }
-
-  if (auto *yield = std::get_if<ast::ExprYield>(&expr.kind)) {
-    if (currentCoroPromisePtr) {
-      // Coroutine-based generator: store value to promise ptr and call suspend marker
-      if (yield->value.has_value() && *yield->value) {
-        auto prevInsideGeneratorYieldValue = insideGeneratorYieldValue;
-        insideGeneratorYieldValue = true;
-        auto yieldVal = generateExpression((*yield->value)->value);
-        insideGeneratorYieldValue = prevInsideGeneratorYieldValue;
-        if (yieldVal) {
-          auto yieldLocation = currentLoc;
-          auto ptrTy = mlir::LLVM::LLVMPointerType::get(&context);
-
-          // Store the yielded value into the promise slot
-          mlir::LLVM::StoreOp::create(builder, yieldLocation, yieldVal, currentCoroPromisePtr);
-
-          // Call the suspend marker: __hew_coro_suspend(promise_ptr)
-          auto suspendMarker = module.lookupSymbol<mlir::func::FuncOp>("__hew_coro_suspend");
-          if (suspendMarker) {
-            mlir::func::CallOp::create(builder, yieldLocation, suspendMarker,
-                                       mlir::ValueRange{currentCoroPromisePtr});
-          }
-
-          return yieldVal; // Return the yielded value (unused by caller)
-        }
-      }
-      return nullptr;
-    }
-    if (currentGenCtx) {
-      // Thread-based generator: call hew_gen_yield(ctx, &val, sizeof(val))
-      if (yield->value.has_value() && *yield->value) {
-        auto prevInsideGeneratorYieldValue = insideGeneratorYieldValue;
-        insideGeneratorYieldValue = true;
-        auto yieldVal = generateExpression((*yield->value)->value);
-        insideGeneratorYieldValue = prevInsideGeneratorYieldValue;
-        if (yieldVal) {
-          auto yieldLocation = currentLoc;
-          auto ptrTy = mlir::LLVM::LLVMPointerType::get(&context);
-          auto i64Ty = builder.getI64Type();
-          auto valType = yieldVal.getType();
-
-          // Alloca for the value
-          auto one = mlir::arith::ConstantIntOp::create(builder, yieldLocation, i64Ty, 1);
-          auto valAlloca =
-              mlir::LLVM::AllocaOp::create(builder, yieldLocation, ptrTy, valType, one);
-          mlir::LLVM::StoreOp::create(builder, yieldLocation, yieldVal, valAlloca);
-
-          // Compute sizeof(valType)
-          auto valSize = hew::SizeOfOp::create(builder, yieldLocation, sizeType(),
-                                               mlir::TypeAttr::get(valType));
-
-          // Call hew_gen_yield(ctx, &val, size) -> bool (i1)
-          auto i1Ty = builder.getI1Type();
-          hew::GenYieldOp::create(builder, yieldLocation, i1Ty, currentGenCtx, valAlloca, valSize);
-          // For now, ignore the return value (cancellation not handled)
-          return yieldVal; // Return the yielded value (unused by caller)
-        }
-      }
-    }
-    // Yield expressions outside generator context are handled at the
-    // statement level during static generator codegen.
-    if (!currentGenCtx && !currentCoroPromisePtr) {
-      ++errorCount_;
-      emitError(currentLoc) << "yield expression outside generator function";
-    }
-    return nullptr;
-  }
-
-  if (std::get_if<ast::ExprCooperate>(&expr.kind)) {
-    // Cooperative scheduler yield point
-    hew::CooperateOp::create(builder, currentLoc);
-    return createIntConstant(builder, currentLoc, builder.getI32Type(), 0);
-  }
-
-  if (std::get_if<ast::ExprThis>(&expr.kind)) {
-    // Actor self-reference handle — must be inside an actor definition
-    assert(!currentActorName.empty() &&
-           "'this' expression encountered outside of an actor context");
-    auto refType = hew::TypedActorRefType::get(&context, builder.getStringAttr(currentActorName));
-    return hew::ActorSelfOp::create(builder, currentLoc, refType).getResult();
-  }
-
-  if (auto *fa = std::get_if<ast::ExprFieldAccess>(&expr.kind)) {
-    auto location = currentLoc;
-    if (insideGeneratorYieldValue) {
-      ++errorCount_;
-      emitError(location) << "yielding a field from a generator is not yet supported "
-                          << "(field-alias ownership tracking is required); "
-                          << "yield the whole value or clone the field instead";
-      return nullptr;
-    }
-    auto operandVal = generateExpression(fa->object->value);
-    if (!operandVal)
-      return nullptr;
-
-    auto operandType = operandVal.getType();
-    const auto &fieldName = fa->field;
-
-    // Handle pointer operands (e.g., `self` in actor receive handlers)
-    if (isPointerLikeType(operandType)) {
-      // Check for handle type property access (e.g., req.path, req.method)
-      if (auto handleTy = mlir::dyn_cast<hew::HandleType>(operandVal.getType())) {
-        auto strRefType = hew::StringRefType::get(&context);
-        if (handleTy.getHandleKind() == "http.Request") {
-          if (fieldName == "path") {
-            return hew::RuntimeCallOp::create(
-                       builder, location, mlir::TypeRange{strRefType},
-                       mlir::SymbolRefAttr::get(&context, "hew_http_request_path"),
-                       mlir::ValueRange{operandVal})
-                .getResult();
-          }
-          if (fieldName == "method") {
-            return hew::RuntimeCallOp::create(
-                       builder, location, mlir::TypeRange{strRefType},
-                       mlir::SymbolRefAttr::get(&context, "hew_http_request_method"),
-                       mlir::ValueRange{operandVal})
-                .getResult();
-          }
-        }
-      }
-
-      // Named supervisor child access: sup.child_name → supervisor_child(sup, idx)
-      if (auto *objIdent = std::get_if<ast::ExprIdentifier>(&fa->object->value.kind)) {
-        auto avIt = actorVarTypes.find(objIdent->name);
-        if (avIt != actorVarTypes.end()) {
-          auto scnIt = supervisorChildNames.find(avIt->second);
-          if (scnIt != supervisorChildNames.end()) {
-            const auto &childNameTypes = scnIt->second;
-            for (size_t i = 0; i < childNameTypes.size(); ++i) {
-              if (childNameTypes[i].first == fieldName) {
-                auto ptrType = mlir::LLVM::LLVMPointerType::get(&context);
-                auto i32Type = builder.getI32Type();
-                auto idxVal =
-                    createIntConstant(builder, location, i32Type, static_cast<int64_t>(i));
-
-                // Check if the child is itself a supervisor
-                bool childIsSupervisor = supervisorChildren.count(childNameTypes[i].second) > 0;
-                if (childIsSupervisor) {
-                  return hew::RuntimeCallOp::create(
-                             builder, location, mlir::TypeRange{ptrType},
-                             mlir::SymbolRefAttr::get(&context,
-                                                      "hew_supervisor_get_child_supervisor"),
-                             mlir::ValueRange{operandVal, idxVal})
-                      .getResult();
-                }
-                // Wait up to 5 seconds for the child to be available
-                // (handles restart windows where children[idx] is temporarily null).
-                auto timeoutVal = mlir::arith::ConstantIntOp::create(builder, location,
-                                                                     builder.getI32Type(), 5000);
-                return hew::RuntimeCallOp::create(
-                           builder, location, mlir::TypeRange{ptrType},
-                           mlir::SymbolRefAttr::get(&context, "hew_supervisor_get_child_wait"),
-                           mlir::ValueRange{operandVal, idxVal, timeoutVal})
-                    .getResult();
-              }
-            }
-          }
-        }
-      }
-
-      std::string targetStructName;
-
-      // Search structTypes — prefer the specific struct when known
-      for (const auto &[typeName, stInfo] : structTypes) {
-        if (!targetStructName.empty() && typeName != targetStructName)
-          continue;
-        auto structType = mlir::dyn_cast<mlir::LLVM::LLVMStructType>(stInfo.mlirType);
-        if (!structType)
-          continue;
-
-        for (const auto &field : stInfo.fields) {
-          if (field.name == fieldName) {
-            auto ptrType = mlir::LLVM::LLVMPointerType::get(&context);
-            auto fieldPtr = mlir::LLVM::GEPOp::create(
-                builder, location, ptrType, structType, operandVal,
-                llvm::ArrayRef<mlir::LLVM::GEPArg>{0, static_cast<int32_t>(field.index)});
-            auto fieldVal =
-                mlir::LLVM::LoadOp::create(builder, location, field.type, fieldPtr).getResult();
-            if (field.semanticType != field.type)
-              return coerceType(fieldVal, field.semanticType, location);
-            return fieldVal;
-          }
-        }
-      }
-      ++errorCount_;
-      emitError(location) << "field '" << fieldName << "' not found on pointer type";
-      return nullptr;
-    }
-
-    // Check for Hew tuple type (numeric field access: t.0, t.1)
-    if (auto hewTuple = mlir::dyn_cast<hew::HewTupleType>(operandType)) {
-      char *end = nullptr;
-      unsigned long numericIdx = std::strtoul(fieldName.c_str(), &end, 10);
-      bool isNumericField = (end != fieldName.c_str() && *end == '\0');
-      if (!isNumericField) {
-        ++errorCount_;
-        emitError(location) << "named field access on tuple type";
-        return nullptr;
-      }
-      auto elemTypes = hewTuple.getElementTypes();
-      if (numericIdx >= elemTypes.size()) {
-        ++errorCount_;
-        emitError(location) << "tuple index " << numericIdx << " out of bounds (size "
-                            << elemTypes.size() << ")";
-        return nullptr;
-      }
-      return hew::TupleExtractOp::create(builder, location, elemTypes[numericIdx], operandVal,
-                                         numericIdx);
-    }
-
-    auto structType = mlir::dyn_cast<mlir::LLVM::LLVMStructType>(operandType);
-    if (!structType) {
-      ++errorCount_;
-      emitError(location) << "field access on non-struct type";
-      return nullptr;
-    }
-
-    // Check for numeric field name (tuple element access: t.0, t.1, etc.)
-    char *end = nullptr;
-    unsigned long numericIdx = std::strtoul(fieldName.c_str(), &end, 10);
-    bool isNumericField = (end != fieldName.c_str() && *end == '\0');
-
-    if (isNumericField) {
-      // Direct numeric index into struct/tuple
-      auto bodyTypes = structType.getBody();
-      if (numericIdx >= bodyTypes.size()) {
-        ++errorCount_;
-        emitError(location) << "tuple index " << numericIdx << " out of bounds (size "
-                            << bodyTypes.size() << ")";
-        return nullptr;
-      }
-      return mlir::LLVM::ExtractValueOp::create(builder, location, operandVal, numericIdx);
-    }
-
-    // Named field: look up struct info by type name
-    // Special case: event types in machine transitions use anonymous structs.
-    // Try enum-based resolution first for event field access.
-    if (!structType.isIdentified() && !currentMachineEventTypeName_.empty()) {
-      auto enumIt = enumTypes.find(currentMachineEventTypeName_);
-      if (enumIt != enumTypes.end() && !currentMachineEventVariant_.empty()) {
-        for (const auto &variant : enumIt->second.variants) {
-          if (variant.name != currentMachineEventVariant_)
-            continue;
-          for (size_t i = 0; i < variant.fieldNames.size(); ++i) {
-            if (variant.fieldNames[i] == fieldName) {
-              auto fieldTy = variant.payloadTypes[i];
-              return hew::EnumExtractPayloadOp::create(builder, location, fieldTy, operandVal,
-                                                       variant.payloadPositions[i]);
-            }
-          }
-          break;
-        }
-      }
-    }
-    if (!structType.isIdentified()) {
-      ++errorCount_;
-      emitError(location) << "named field access on anonymous struct type";
-      return nullptr;
-    }
-    llvm::StringRef structName = structType.getName();
-    auto it = structTypes.find(structName.str());
-    if (it != structTypes.end()) {
-      const auto &info = it->second;
-
-      // Find the field
-      for (const auto &field : info.fields) {
-        if (field.name == fieldName) {
-          auto fieldVal = hew::FieldGetOp::create(builder, location, field.type, operandVal,
-                                                  builder.getStringAttr(fieldName),
-                                                  builder.getI64IntegerAttr(field.index))
-                              .getResult();
-          if (field.semanticType != field.type)
-            return coerceType(fieldVal, field.semanticType, location);
-          return fieldVal;
-        }
-      }
-
-      {
-        ++errorCount_;
-        auto diag = emitError(location)
-                    << "no field '" << fieldName << "' on struct '" << structName << "'";
-        if (!info.fields.empty()) {
-          diag << "; available fields: ";
-          for (size_t i = 0; i < info.fields.size(); ++i) {
-            if (i > 0)
-              diag << ", ";
-            diag << info.fields[i].name;
-          }
-        }
-      }
-      return nullptr;
-    }
-
-    // Machine/enum field access: resolve `state.field` and `event.field` inside transition bodies
-    std::string lookupName = structName.str();
-    // Event types use anonymous LLVM structs — look up by registered event type name instead
-    if (lookupName.empty() && !currentMachineEventTypeName_.empty()) {
-      lookupName = currentMachineEventTypeName_;
-    }
-    auto enumIt = enumTypes.find(lookupName);
-    if (enumIt != enumTypes.end() && !currentMachineSourceVariant_.empty()) {
-      const auto &enumInfo = enumIt->second;
-      // Determine which variant to resolve: for event types use the event variant,
-      // otherwise use the source state variant (for `state.field`).
-      const std::string &variantName = (!currentMachineEventTypeName_.empty() &&
-                                        structName.str() == currentMachineEventTypeName_)
-                                           ? currentMachineEventVariant_
-                                           : currentMachineSourceVariant_;
-      for (const auto &variant : enumInfo.variants) {
-        if (variant.name != variantName)
-          continue;
-        for (size_t i = 0; i < variant.fieldNames.size(); ++i) {
-          if (variant.fieldNames[i] == fieldName) {
-            auto fieldTy = getEnumFieldType(operandType, variant.payloadPositions[i]);
-            return hew::EnumExtractPayloadOp::create(builder, location, fieldTy, operandVal,
-                                                     variant.payloadPositions[i]);
-          }
-        }
-        break;
-      }
-    }
-
-    ++errorCount_;
-    emitError(location) << "unknown struct type '" << structName << "'";
-    return nullptr;
-  }
-
-  if (auto *idx = std::get_if<ast::ExprIndex>(&expr.kind)) {
-    auto location = currentLoc;
-    auto operandVal = generateExpression(idx->object->value);
-    if (!operandVal)
-      return nullptr;
-
-    // Hew array type indexing
-    if (auto hewArrayType = mlir::dyn_cast<hew::HewArrayType>(operandVal.getType())) {
-      auto indexVal = generateExpression(idx->index->value);
-      if (!indexVal)
-        return nullptr;
-
-      // Constant index -> hew.array_extract
-      if (auto constOp = indexVal.getDefiningOp<mlir::arith::ConstantIntOp>()) {
-        auto idxConst = constOp.value();
-        return hew::ArrayExtractOp::create(builder, location, hewArrayType.getElementType(),
-                                           operandVal, idxConst);
-      }
-
-      // Dynamic index -> cast to LLVM array, spill to alloca, GEP, load
-      auto llvmArrayType =
-          mlir::LLVM::LLVMArrayType::get(hewArrayType.getElementType(), hewArrayType.getSize());
-      auto llvmArray = hew::BitcastOp::create(builder, location, llvmArrayType, operandVal);
-      auto alloca = mlir::LLVM::AllocaOp::create(
-          builder, location, mlir::LLVM::LLVMPointerType::get(&context), llvmArrayType,
-          mlir::arith::ConstantIntOp::create(builder, location, 1, 64));
-      mlir::LLVM::StoreOp::create(builder, location, llvmArray, alloca);
-      auto i64Type = builder.getI64Type();
-      mlir::Value idx64 = indexVal;
-      if (indexVal.getType() != i64Type)
-        idx64 = mlir::arith::ExtSIOp::create(builder, location, i64Type, indexVal);
-      auto zero = mlir::arith::ConstantIntOp::create(builder, location, 0, 64);
-      auto elemPtr =
-          mlir::LLVM::GEPOp::create(builder, location, mlir::LLVM::LLVMPointerType::get(&context),
-                                    llvmArrayType, alloca, mlir::ValueRange{zero, idx64});
-      return mlir::LLVM::LoadOp::create(builder, location, hewArrayType.getElementType(), elemPtr);
-    }
-
-    auto arrayType = mlir::dyn_cast<mlir::LLVM::LLVMArrayType>(operandVal.getType());
-    if (arrayType) {
-      auto indexVal = generateExpression(idx->index->value);
-      if (!indexVal)
-        return nullptr;
-
-      // Constant index -> extractvalue
-      if (auto constOp = indexVal.getDefiningOp<mlir::arith::ConstantIntOp>()) {
-        auto idxConst = constOp.value();
-        return mlir::LLVM::ExtractValueOp::create(builder, location, operandVal, idxConst);
-      }
-
-      // Dynamic index -> spill array to alloca, GEP, load
-      auto alloca = mlir::LLVM::AllocaOp::create(
-          builder, location, mlir::LLVM::LLVMPointerType::get(&context), arrayType,
-          mlir::arith::ConstantIntOp::create(builder, location, 1, 64));
-      mlir::LLVM::StoreOp::create(builder, location, operandVal, alloca);
-      auto i64Type = builder.getI64Type();
-      mlir::Value idx64 = indexVal;
-      if (indexVal.getType() != i64Type)
-        idx64 = mlir::arith::ExtSIOp::create(builder, location, i64Type, indexVal);
-      auto zero = mlir::arith::ConstantIntOp::create(builder, location, 0, 64);
-      auto elemPtr =
-          mlir::LLVM::GEPOp::create(builder, location, mlir::LLVM::LLVMPointerType::get(&context),
-                                    arrayType, alloca, mlir::ValueRange{zero, idx64});
-      return mlir::LLVM::LoadOp::create(builder, location, arrayType.getElementType(), elemPtr);
-    }
-
-    if (auto vecType = mlir::dyn_cast<hew::VecType>(operandVal.getType())) {
-      auto indexVal = generateExpression(idx->index->value);
-      if (!indexVal)
-        return nullptr;
-      auto i64Type = builder.getI64Type();
-      mlir::Value idx64 = indexVal;
-      if (indexVal.getType() != i64Type)
-        idx64 = mlir::arith::ExtSIOp::create(builder, location, i64Type, indexVal);
-      return hew::VecGetOp::create(builder, location, vecType.getElementType(), operandVal, idx64);
-    }
-
-    // Custom type indexing: desugar obj[key] → obj.get(key)
-    if (auto structType = mlir::dyn_cast<mlir::LLVM::LLVMStructType>(operandVal.getType())) {
-      if (structType.isIdentified()) {
-        auto indexVal = generateExpression(idx->index->value);
-        if (!indexVal)
-          return nullptr;
-
-        std::string funcName = mangleName(currentModulePath, structType.getName().str(), "get");
-        llvm::SmallVector<mlir::Value, 2> args;
-        args.push_back(operandVal);
-        args.push_back(indexVal);
-
-        auto callee = module.lookupSymbol<mlir::func::FuncOp>(funcName);
-        if (!callee)
-          callee = lookupImportedFunc(structType.getName(), "get");
-        if (callee) {
-          auto funcType = callee.getFunctionType();
-          for (size_t i = 0; i < args.size() && i < funcType.getNumInputs(); ++i) {
-            if (args[i].getType() != funcType.getInput(i)) {
-              args[i] = coerceType(args[i], funcType.getInput(i), location);
-              if (!args[i])
-                return nullptr;
-            }
-          }
-          auto callOp = mlir::func::CallOp::create(builder, location, callee, args);
-          if (callOp.getNumResults() > 0)
-            return callOp.getResult(0);
-          return nullptr;
-        }
-      }
-    }
-
-    ++errorCount_;
-    emitError(location) << "indexing not supported for this type";
-    return nullptr;
-  }
-
-  if (auto *awaitE = std::get_if<ast::ExprAwait>(&expr.kind)) {
-    auto location = currentLoc;
-    // Check if the inner expression is a method call (actor ask pattern)
-    if (auto *mc = std::get_if<ast::ExprMethodCall>(&awaitE->inner->value.kind)) {
-      auto mcLocation = loc(awaitE->inner->span);
-
-      // Generate receiver
-      auto receiver = generateExpression(mc->receiver->value);
-      if (!receiver)
-        return nullptr;
-
-      // Resolve actor type
-      std::string actorTypeName = resolveActorTypeName(mc->receiver->value, &mc->receiver->span);
-
-      if (!actorTypeName.empty()) {
-        auto actorIt = actorRegistry.find(actorTypeName);
-        if (actorIt != actorRegistry.end()) {
-          return generateActorMethodAsk(receiver, actorIt->second, mc->method, mc->args,
-                                        mcLocation);
-        }
-      }
-
-      ++errorCount_;
-      emitError(location) << "await requires an actor method call";
-      return nullptr;
-    }
-    // Not a method call — operand might be an actor ref or a task handle.
-    auto operand = generateExpression(awaitE->inner->value);
-    if (!operand)
-      return nullptr;
-
-    if (mlir::isa<hew::ActorRefType, hew::TypedActorRefType>(operand.getType())) {
-      auto awaitOp = builder.create<hew::ActorAwaitOp>(location, builder.getI32Type(), operand);
-      return awaitOp.getResult();
-    }
-
-    auto ptrType = mlir::LLVM::LLVMPointerType::get(&context);
-    if (auto handleTy = mlir::dyn_cast<hew::HandleType>(operand.getType())) {
-      if (handleTy.getHandleKind() == "Task")
-        operand = hew::BitcastOp::create(builder, location, ptrType, operand);
-    }
-    if (mlir::isa<mlir::LLVM::LLVMPointerType>(operand.getType())) {
-      auto resultPtr = hew::ScopeAwaitOp::create(builder, location, ptrType, operand);
-
-      auto resolveScopeAwaitResultType = [&]() -> mlir::Type {
-        if (auto *ie = std::get_if<ast::ExprIdentifier>(&awaitE->inner->value.kind)) {
-          auto it = taskResultTypes.find(ie->name);
-          if (it != taskResultTypes.end())
-            return it->second;
-        }
-
-        const auto *resolvedTaskType =
-            requireResolvedTypeOf(awaitE->inner->span, "scope.await operand", location);
-        if (!resolvedTaskType)
-          return nullptr;
-
-        auto *taskNamed = std::get_if<ast::TypeNamed>(&resolvedTaskType->kind);
-        if (!taskNamed) {
-          ++errorCount_;
-          emitError(location) << "scope.await operand must resolve to Task<T>";
-          return nullptr;
-        }
-
-        if ((!taskNamed->type_args || taskNamed->type_args->size() != 1) &&
-            resolveTypeAlias(taskNamed->name) != taskNamed->name) {
-          if (const auto *aliasType = resolveTypeAliasExpr(taskNamed->name)) {
-            if (const auto *aliasNamed = std::get_if<ast::TypeNamed>(&aliasType->kind))
-              taskNamed = aliasNamed;
-          }
-        }
-
-        auto taskName = resolveTypeAlias(taskNamed->name);
-        if ((taskName != "Task" && taskName != "scope.Task") ||
-            !(taskNamed->type_args && taskNamed->type_args->size() == 1)) {
-          ++errorCount_;
-          emitError(location) << "scope.await operand must resolve to Task<T>";
-          return nullptr;
-        }
-
-        return convertTypeOrError((*taskNamed->type_args)[0].value,
-                                  "cannot determine scope.await result type", location);
-      };
-
-      auto resultType = resolveScopeAwaitResultType();
-      if (!resultType)
-        return nullptr;
-
-      auto loadedResult = mlir::LLVM::LoadOp::create(builder, location, resultType, resultPtr);
-      return loadedResult;
-    }
-    return operand;
-  }
-
-  if (auto *range = std::get_if<ast::ExprRange>(&expr.kind)) {
-    // Range expression: start..end or start..=end
-    if (!range->start || !range->end) {
-      ++errorCount_;
-      emitError(currentLoc) << "unbounded ranges not yet supported as values";
-      return nullptr;
-    }
-
-    auto startVal = generateExpression((*range->start)->value);
-    auto endVal = generateExpression((*range->end)->value);
-
-    if (!startVal || !endVal)
-      return nullptr;
-
-    // Ensure types match
-    if (startVal.getType() != endVal.getType()) {
-      endVal = coerceType(endVal, startVal.getType(), currentLoc);
-      if (!endVal)
-        return nullptr;
-    }
-
-    if (range->inclusive) {
-      // For inclusive range ..=, add 1 to end value (assuming integer)
-      if (endVal.getType().isIntOrIndex()) {
-        auto one = createIntConstant(builder, currentLoc, endVal.getType(), 1);
-        endVal = mlir::arith::AddIOp::create(builder, currentLoc, endVal, one);
-      } else {
-        ++errorCount_;
-        emitError(currentLoc) << "inclusive range only supported for integers";
-        return nullptr;
-      }
-    }
-
-    auto tupleType = hew::HewTupleType::get(&context, {startVal.getType(), endVal.getType()});
-    return hew::TupleCreateOp::create(builder, currentLoc, tupleType,
-                                      mlir::ValueRange{startVal, endVal});
-  }
-
-  ++errorCount_;
-  emitError(currentLoc) << "unsupported expression kind";
-  return nullptr;
-}
-
-// ============================================================================
 // Interpolated string generation
 // ============================================================================
