+++ hew-codegen/src/mlir/MLIRGenStmt.cpp
@@ -507,75 +507,6 @@ mlir::Value MLIRGen::generateBlock(const ast::Block &block, bool statementPositi
 // Statement generation
 // ============================================================================
 
-void MLIRGen::generateStatement(const ast::Stmt &stmt) {
-  currentLoc = loc(stmt.span);
-  if (auto *s = std::get_if<ast::StmtLet>(&stmt.kind)) {
-    generateLetStmt(*s);
-    return;
-  }
-  if (auto *s = std::get_if<ast::StmtVar>(&stmt.kind)) {
-    generateVarStmt(*s);
-    return;
-  }
-  if (auto *s = std::get_if<ast::StmtAssign>(&stmt.kind)) {
-    generateAssignStmt(*s);
-    return;
-  }
-  if (auto *s = std::get_if<ast::StmtIf>(&stmt.kind)) {
-    generateIfStmt(*s);
-    return;
-  }
-  if (auto *s = std::get_if<ast::StmtWhile>(&stmt.kind)) {
-    generateWhileStmt(*s);
-    return;
-  }
-  if (auto *s = std::get_if<ast::StmtWhileLet>(&stmt.kind)) {
-    generateWhileLetStmt(*s);
-    return;
-  }
-  if (auto *s = std::get_if<ast::StmtFor>(&stmt.kind)) {
-    if (s->is_await)
-      generateForAwaitStmt(*s);
-    else
-      generateForStmt(*s);
-    return;
-  }
-  if (auto *s = std::get_if<ast::StmtReturn>(&stmt.kind)) {
-    generateReturnStmt(*s);
-    return;
-  }
-  if (auto *s = std::get_if<ast::StmtExpression>(&stmt.kind)) {
-    generateExprStmt(*s);
-    return;
-  }
-  if (auto *s = std::get_if<ast::StmtLoop>(&stmt.kind)) {
-    generateLoopStmt(*s);
-    return;
-  }
-  if (auto *s = std::get_if<ast::StmtMatch>(&stmt.kind)) {
-    generateMatchStmt(*s);
-    return;
-  }
-  if (auto *s = std::get_if<ast::StmtBreak>(&stmt.kind)) {
-    generateBreakStmt(*s);
-    return;
-  }
-  if (auto *s = std::get_if<ast::StmtContinue>(&stmt.kind)) {
-    generateContinueStmt(*s);
-    return;
-  }
-  if (auto *s = std::get_if<ast::StmtDefer>(&stmt.kind)) {
-    if (s->expr) {
-      currentFnDefers.emplace_back(&s->expr->value, currentLoc);
-    }
-    return;
-  }
-  if (auto *ifLet = std::get_if<ast::StmtIfLet>(&stmt.kind)) {
-    generateIfLetStmt(*ifLet);
-    return;
-  }
-}
-
 void MLIRGen::bindLetSubPattern(const ast::Pattern &pattern, mlir::Value value,
                                 mlir::Location location) {
   if (auto *identPat = std::get_if<ast::PatIdentifier>(&pattern.kind)) {
